Split accuracy unit test for convolution operators
