+++ tests/test_conv1d.py
@@ -1,0 +1,102 @@
+import pytest
+import torch
+
+import flag_gems
+
+from . import accuracy_utils as utils
+
+SHAPE_CONV1D = [
+    ((32, 2, 4), (17, 2, 2)),
+    ((32, 15, 6), (17, 15, 2)),
+    ((64, 64, 64), (128, 64, 7)),
+    # ((32, 16, 1024), (1024, 16, 8)),
+    # ((32, 12, 9), (17, 12, 3)),
+    # ((32, 6, 6), (64, 6, 2)),
+]
+
+SHAPE_CONV1D_DILATION = [
+    ((32, 2, 16), (17, 2, 3)),
+    ((32, 15, 32), (17, 15, 3)),
+    ((64, 64, 64), (128, 64, 3)),
+]
+
+
+@pytest.mark.conv1d
+@pytest.mark.parametrize("shape, kernel", SHAPE_CONV1D)
+@pytest.mark.parametrize("stride", [2])
+@pytest.mark.parametrize("padding", [1])
+@pytest.mark.parametrize("dtype", [torch.float32, torch.float16])
+def test_conv1d(monkeypatch, shape, kernel, stride, padding, dtype):
+    if flag_gems.vendor_name == "mthreads" and dtype == torch.float16:
+        monkeypatch.env("MUSA_ENABLE_SQMMA", "1")
+
+    inp = torch.randn(shape, dtype=dtype, device=flag_gems.device, requires_grad=True)
+    ref_inp = utils.to_reference(inp, True)
+    weight = torch.randn(kernel, dtype=dtype, device=flag_gems.device)
+    ref_weight = utils.to_reference(weight, True)
+    ref_out = torch.nn.functional.conv1d(
+        ref_inp, ref_weight, bias=None, stride=stride, padding=padding, dilation=1
+    )
+
+    res_out = flag_gems.conv1d(
+        inp, weight, bias=None, stride=stride, padding=padding, dilation=1
+    )
+    utils.gems_assert_close(res_out, ref_out, dtype)
+
+
+@pytest.mark.conv1d_padding
+@pytest.mark.skipif(flag_gems.vendor_name == "kunlunxin", reason="RESULT TODOFIX")
+@pytest.mark.parametrize("shape, kernel", SHAPE_CONV1D)
+@pytest.mark.parametrize("stride", [1])
+@pytest.mark.parametrize("padding", ["valid", "same"])
+@pytest.mark.parametrize("dtype", [torch.float32, torch.float16])
+def test_accuracy_conv1d_padding(monkeypatch, shape, kernel, stride, padding, dtype):
+    if flag_gems.vendor_name == "mthreads" and dtype == torch.float16:
+        monkeypatch.env("MUSA_ENABLE_SQMMA", "1")
+
+    inp = torch.randn(shape, dtype=dtype, device=flag_gems.device, requires_grad=True)
+    ref_inp = utils.to_reference(inp, True)
+    weight = torch.randn(kernel, dtype=dtype, device=flag_gems.device)
+    ref_weight = utils.to_reference(weight, True)
+    ref_out = torch.nn.functional.conv1d(
+        ref_inp, ref_weight, bias=None, stride=stride, padding=padding, dilation=1
+    )
+
+    res_out = flag_gems.conv1d(
+        inp, weight, bias=None, stride=stride, padding=padding, dilation=1
+    )
+    utils.gems_assert_close(res_out, ref_out, dtype)
+
+
+@pytest.mark.conv1d
+@pytest.mark.parametrize("shape, kernel", SHAPE_CONV1D_DILATION)
+@pytest.mark.parametrize("stride", [1])
+@pytest.mark.parametrize("padding", [0, 2])
+@pytest.mark.parametrize("dtype", [torch.float32, torch.float16])
+@pytest.mark.parametrize("dilation", [1, 2, (1,), (2,)])
+def test_accuracy_conv1d_dilation(shape, kernel, stride, padding, dtype, dilation):
+    """Test conv1d with various dilation values, including tuple form.
+
+    This specifically tests the fix where conv1d must properly convert dilation
+    to a 2D tuple before delegating to conv2d. Previously, passing dilation as
+    a single-element tuple (e.g., (1,)) would cause a ValueError in conv2d.
+    """
+    inp = torch.randn(shape, dtype=dtype, device=flag_gems.device, requires_grad=True)
+    ref_inp = utils.to_reference(inp, True)
+    weight = torch.randn(kernel, dtype=dtype, device=flag_gems.device)
+    ref_weight = utils.to_reference(weight, True)
+
+    ref_out = torch.nn.functional.conv1d(
+        ref_inp,
+        ref_weight,
+        bias=None,
+        stride=stride,
+        padding=padding,
+        dilation=dilation,
+    )
+
+    res_out = flag_gems.conv1d(
+        inp, weight, bias=None, stride=stride, padding=padding, dilation=dilation
+    )
+
+    utils.gems_assert_close(res_out, ref_out, dtype)
